imap: move over imapclient.NamespaceData
